wip on wasm test for zome get
